+++ src/common/commands.ts
@@ -1,12 +1,18 @@
+import { AsyncLocalStorage } from "node:async_hooks";
+import * as crypto from "node:crypto";
+import * as events from "node:events";
 import * as vscode from "vscode";
 import type { BuildManager } from "../build/manager";
 import type { DestinationsManager } from "../destination/manager";
 import type { DestinationType, SelectedDestination } from "../destination/types";
+import type { SwiftFormattingProvider } from "../format/formatter";
+import type { ProgressStatusBar } from "../system/status-bar";
 import type { TestingManager } from "../testing/manager";
 import type { ToolsManager } from "../tools/manager";
 import { addTreeProviderErrorReporting, errorReporting } from "./error-reporting";
 import { type ErrorMessageAction, ExtensionError, TaskError } from "./errors";
 import { commonLogger } from "./logger";
+import { QuickPickCancelledError } from "./quick-pick";
 
 export type LastLaunchedAppDeviceContext = {
   type: "device";
@@ -51,13 +57,29 @@ type WorkspaceTypes = {
 type WorkspaceStateKey = keyof WorkspaceTypes;
 type SessionStateKey = "NONE_KEY";
 
+/**
+ * Global events that extension can emit
+ */
+type IEventMap = {
+  executionScopeClosed: [scope: ExecutionScope];
+  workspaceConfigChanged: [];
+};
+type IEventKey = keyof IEventMap;
+
 export class ExtensionContext {
   private _context: vscode.ExtensionContext;
   public destinationsManager: DestinationsManager;
   public toolsManager: ToolsManager;
   public buildManager: BuildManager;
   public testingManager: TestingManager;
+  public formatter: SwiftFormattingProvider;
+  public progressStatusBar: ProgressStatusBar;
   private _sessionState: Map<SessionStateKey, unknown> = new Map();
+
+  // Create for each command and task execution separate execution scope with unique ID
+  // to be able to track what is currently running
+  private executionScope = new AsyncLocalStorage<ExecutionScope | undefined>();
+  private emitter = new events.EventEmitter<IEventMap>();
 
   constructor(options: {
     context: vscode.ExtensionContext;
@@ -65,12 +87,23 @@ export class ExtensionContext {
     buildManager: BuildManager;
     toolsManager: ToolsManager;
     testingManager: TestingManager;
+    formatter: SwiftFormattingProvider;
+    progressStatusBar: ProgressStatusBar;
   }) {
     this._context = options.context;
     this.destinationsManager = options.destinationsManager;
     this.buildManager = options.buildManager;
     this.toolsManager = options.toolsManager;
     this.testingManager = options.testingManager;
+    this.formatter = options.formatter;
+    this.progressStatusBar = options.progressStatusBar;
+
+    vscode.workspace.onDidChangeConfiguration((event) => {
+      const affected = event.affectsConfiguration("sweetpad");
+      if (affected) {
+        this.emitter.emit("workspaceConfigChanged");
+      }
+    });
   }
 
   get storageUri() {
@@ -85,11 +118,119 @@ export class ExtensionContext {
     this._context.subscriptions.push(disposable);
   }
 
-  registerCommand(command: string, callback: (context: CommandExecution, ...args: any[]) => Promise<unknown>) {
-    return vscode.commands.registerCommand(command, (...args: any[]) => {
-      const execution = new CommandExecution(command, callback, this);
-      return execution.run(...args);
+  /**
+   * In case if you need to start propage execution scope manually you can use this method
+   */
+  setExecutionScope<T>(scope: ExecutionScope | undefined, callback: () => Promise<T>): Promise<T> {
+    return this.executionScope.run(scope, callback);
+  }
+
+  getExecutionScope(): ExecutionScope | undefined {
+    return this.executionScope.getStore();
+  }
+
+  getExecutionScopeId(): string | undefined {
+    return this.getExecutionScope()?.id;
+  }
+
+  /**
+   * Main method to start execution scope for command or task or other isolated execution context
+   */
+  startExecutionScope<T>(scope: ExecutionScope, callback: () => Promise<T>): Promise<T> {
+    return this.executionScope.run(scope, async () => {
+      try {
+        return await callback();
+      } finally {
+        this.emitter.emit("executionScopeClosed", scope);
+      }
     });
+  }
+
+  on<K extends IEventKey>(event: K, listener: (...args: IEventMap[K]) => void): void {
+    this.emitter.on(event, listener as any); // todo: fix this any
+  }
+
+  registerCommand(commandName: string, callback: (context: ExtensionContext, ...args: any[]) => Promise<unknown>) {
+    return vscode.commands.registerCommand(commandName, async (...args: any[]) => {
+      const commandContext = new CommandExecutionScope({ commandName: commandName });
+
+      return await errorReporting.withScope(async (scope) => {
+        return await this.startExecutionScope(commandContext, async () => {
+          scope.setTag("command", commandName);
+          try {
+            return await callback(this, ...args);
+          } catch (error) {
+            // User can cancel the quick pick dialog by pressing Escape or clicking outside of it.
+            // In this case, we just stop the execution of the command and throw a QuickPickCancelledError.
+            // Since it is more user action, then an error, we skip the error reporting.
+            if (error instanceof QuickPickCancelledError) {
+              return;
+            }
+
+            if (error instanceof ExtensionError) {
+              // Handle default error
+              commonLogger.error(error.message, {
+                command: commandName,
+                errorContext: error.options?.context,
+                error: error,
+              });
+              if (error instanceof TaskError) {
+                return; // do nothing
+              }
+
+              await this.showCommandErrorMessage(`SweetPad: ${error.message}`, {
+                actions: error.options?.actions,
+              });
+              return;
+            }
+
+            // Handle unexpected error
+            const errorMessage: string =
+              error instanceof Error ? error.message : (error?.toString() ?? "[unknown error]");
+            commonLogger.error(errorMessage, {
+              command: commandName,
+              error: error,
+            });
+            errorReporting.captureException(error);
+            await this.showCommandErrorMessage(`SweetPad: ${errorMessage}`);
+          }
+        });
+      });
+    });
+  }
+
+  /**
+   * Show error message with proper actions
+   */
+  async showCommandErrorMessage(
+    message: string,
+    options?: {
+      actions?: ErrorMessageAction[];
+    },
+  ): Promise<void> {
+    const closeAction: ErrorMessageAction = {
+      label: "Close",
+      callback: () => {},
+    };
+    const showLogsAction: ErrorMessageAction = {
+      label: "Show logs",
+      callback: () => commonLogger.show(),
+    };
+
+    const actions = [closeAction];
+    actions.unshift(...(options?.actions ?? [showLogsAction]));
+
+    const actionsLabels = actions.map((action) => action.label);
+
+    const finalMessage = `${message}`;
+    const action = await vscode.window.showErrorMessage(finalMessage, ...actionsLabels);
+
+    if (action) {
+      const callback = actions.find((a) => a.label === action)?.callback;
+      if (callback) {
+        callback();
+      }
+    }
   }
 
   registerTreeDataProvider<T extends vscode.TreeItem>(id: string, tree: vscode.TreeDataProvider<T>) {
@@ -135,88 +276,34 @@ export class ExtensionContext {
     void this.buildManager.refresh();
     void this.destinationsManager.refresh();
   }
-}
 
-/**
- * Class that represents a command execution with proper error handling
- */
-export class CommandExecution {
-  constructor(
-    public readonly command: string,
-    public readonly callback: (context: CommandExecution, ...args: unknown[]) => Promise<unknown>,
-    public context: ExtensionContext,
-  ) {}
-
-  /**
-   * Show error message with proper actions
-   */
-  async showErrorMessage(
-    message: string,
-    options?: {
-      actions?: ErrorMessageAction[];
-    },
-  ): Promise<void> {
-    const closeAction: ErrorMessageAction = {
-      label: "Close",
-      callback: () => {},
-    };
-    const showLogsAction: ErrorMessageAction = {
-      label: "Show logs",
-      callback: () => commonLogger.show(),
-    };
-
-    const actions = [closeAction];
-    actions.unshift(...(options?.actions ?? [showLogsAction]));
-
-    const actionsLabels = actions.map((action) => action.label);
-
-    const finalMessage = `${message}`;
-    const action = await vscode.window.showErrorMessage(finalMessage, ...actionsLabels);
-
-    if (action) {
-      const callback = actions.find((a) => a.label === action)?.callback;
-      if (callback) {
-        callback();
-      }
-    }
-  }
-
-  /**
-   * Run the command with proper error handling. First argument passed to
-   * the callback is this instance itself.
-   */
-  async run(...args: unknown[]) {
-    return await errorReporting.withScope(async (scope) => {
-      scope.setTag("command", this.command);
-      try {
-        return await this.callback(this, ...args);
-      } catch (error) {
-        if (error instanceof ExtensionError) {
-          // Handle default error
-          commonLogger.error(error.message, {
-            command: this.command,
-            errorContext: error.options?.context,
-            error: error,
-          });
-          if (error instanceof TaskError) {
-            // do nothing
-          } else {
-            await this.showErrorMessage(`Sweetpad: ${error.message}`, {
-              actions: error.options?.actions,
-            });
-          }
-        } else {
-          // Handle unexpected error
-          const errorMessage: string =
-            error instanceof Error ? error.message : (error?.toString() ?? "[unknown error]");
-          commonLogger.error(errorMessage, {
-            command: this.command,
-            error: error,
-          });
-          errorReporting.captureException(error);
-          await this.showErrorMessage(`Sweetpad: ${errorMessage}`);
-        }
-      }
-    });
+  updateProgressStatus(message: string) {
+    this.progressStatusBar.updateText(message);
   }
 }
+
+export class CommandExecutionScope {
+  id: string;
+  type = "command" as const;
+  commandName: string;
+
+  constructor(options: { commandName: string }) {
+    this.id = crypto.randomUUID();
+    this.type = "command";
+    this.commandName = options.commandName;
+  }
+}
+
+export class TaskExecutionScope {
+  id: string;
+  type = "task" as const;
+  taskName: string;
+
+  constructor(options: { action: string }) {
+    this.id = crypto.randomUUID();
+    this.type = "task";
+    this.taskName = options.action;
+  }
+}
+
+export type ExecutionScope = CommandExecutionScope | TaskExecutionScope;