+++ src/extension.ts
@@ -2,6 +2,9 @@ import * as vscode from "vscode";
 import {
   buildCommand,
   cleanCommand,
+  debuggingBuildCommand,
+  debuggingLaunchCommand,
+  debuggingRunCommand,
   diagnoseBuildSetupCommand,
   generateBuildServerConfigCommand,
   launchCommand,
@@ -33,7 +36,7 @@ import { DestinationStatusBar } from "./destination/status-bar.js";
 import { DestinationsTreeProvider } from "./destination/tree.js";
 import { DevicesManager } from "./devices/manager.js";
 import { formatCommand, showLogsCommand } from "./format/commands.js";
-import { createFormatProvider } from "./format/provider.js";
+import { SwiftFormattingProvider, registerFormatProvider, registerRangeFormatProvider } from "./format/formatter.js";
 import { createFormatStatusItem } from "./format/status.js";
 import {
   openSimulatorCommand,
@@ -45,9 +48,11 @@ import { SimulatorsManager } from "./simulators/manager.js";
 import {
   createIssueGenericCommand,
   createIssueNoSchemesCommand,
+  openTerminalPanel,
   resetSweetpadCache,
   testErrorReportingCommand,
 } from "./system/commands.js";
+import { ProgressStatusBar } from "./system/status-bar.js";
 import {
   buildForTestingCommand,
   selectConfigurationForTestingCommand,
@@ -85,6 +90,8 @@ export function activate(context: vscode.ExtensionContext) {
   });
   const toolsManager = new ToolsManager();
   const testingManager = new TestingManager();
+  const formatter = new SwiftFormattingProvider();
+  const progressStatusBar = new ProgressStatusBar();
 
   // Main context object 🌍
   const _context = new ExtensionContext({
@@ -93,12 +100,15 @@ export function activate(context: vscode.ExtensionContext) {
     buildManager: buildManager,
     toolsManager: toolsManager,
     testingManager: testingManager,
+    formatter: formatter,
+    progressStatusBar: progressStatusBar,
   });
   // Here is circular dependency, but I don't care
   buildManager.context = _context;
   devicesManager.context = _context;
   destinationsManager.context = _context;
   testingManager.context = _context;
+  progressStatusBar.context = _context;
 
   // Trees 🎄
   const buildTreeProvider = new BuildTreeProvider({
@@ -119,10 +129,6 @@ export function activate(context: vscode.ExtensionContext) {
 
   const buildTaskProvider = new XcodeBuildTaskProvider(_context);
 
-  // Debug
-  d(registerDebugConfigurationProvider(_context));
-  d(command("sweetpad.debugger.getAppPath", getAppPathCommand));
-
   // Tasks
   d(vscode.tasks.registerTaskProvider(buildTaskProvider.type, buildTaskProvider));
 
@@ -140,7 +146,7 @@ export function activate(context: vscode.ExtensionContext) {
   d(command("sweetpad.build.test", testCommand));
   d(command("sweetpad.build.resolveDependencies", resolveDependenciesCommand));
   d(command("sweetpad.build.removeBundleDir", removeBundleDirCommand));
-  d(command("sweetpad.build.genereateBuildServerConfig", generateBuildServerConfigCommand));
+  d(command("sweetpad.build.generateBuildServerConfig", generateBuildServerConfigCommand));
   d(command("sweetpad.build.openXcode", openXcodeCommand));
   d(command("sweetpad.build.selectXcodeWorkspace", selectXcodeWorkspaceCommand));
   d(command("sweetpad.build.setDefaultScheme", selectXcodeSchemeForBuildCommand));
@@ -153,6 +159,13 @@ export function activate(context: vscode.ExtensionContext) {
   d(command("sweetpad.testing.selectTarget", selectTestingTargetCommand));
   d(command("sweetpad.testing.setDefaultScheme", selectXcodeSchemeForTestingCommand));
   d(command("sweetpad.testing.selectConfiguration", selectConfigurationForTestingCommand));
+
+  // Debugging
+  d(registerDebugConfigurationProvider(_context));
+  d(command("sweetpad.debugger.getAppPath", getAppPathCommand));
+  d(command("sweetpad.debugger.debuggingLaunch", debuggingLaunchCommand));
+  d(command("sweetpad.debugger.debuggingRun", debuggingRunCommand));
+  d(command("sweetpad.debugger.debuggingBuild", debuggingBuildCommand));
 
   // XcodeGen
   d(command("sweetpad.xcodegen.generate", xcodgenGenerateCommand));
@@ -167,7 +180,8 @@ export function activate(context: vscode.ExtensionContext) {
 
   // Format
   d(createFormatStatusItem());
-  d(createFormatProvider());
+  d(registerFormatProvider(formatter));
+  d(registerRangeFormatProvider(formatter));
   d(command("sweetpad.format.run", formatCommand));
   d(command("sweetpad.format.showLogs", showLogsCommand));
 
@@ -202,6 +216,7 @@ export function activate(context: vscode.ExtensionContext) {
   d(command("sweetpad.system.createIssue.generic", createIssueGenericCommand));
   d(command("sweetpad.system.createIssue.noSchemes", createIssueNoSchemesCommand));
   d(command("sweetpad.system.testErrorReporting", testErrorReportingCommand));
+  d(command("sweetpad.system.openTerminalPanel", openTerminalPanel));
 }
 
 export function deactivate() {}