+++ src/format/commands.ts
@@ -1,18 +1,19 @@
 import * as vscode from "vscode";
-import { formatDocument } from "./formatter.js";
+import type { ExtensionContext } from "../common/commands.js";
 import { formatLogger } from "./logger.js";
 
 /*
  * Format current opened document
  */
-export async function formatCommand() {
+export async function formatCommand(context: ExtensionContext) {
+  context.updateProgressStatus("Formatting document");
   const editor = vscode.window.activeTextEditor;
   if (!editor) {
     return;
   }
 
   const document = editor.document;
-  await formatDocument(document);
+  await context.formatter.formatDocument(document);
 }
 
 /*