+++ src/tuist/command.ts
@@ -1,6 +1,7 @@
 import * as vscode from "vscode";
 import { restartSwiftLSP } from "../build/utils";
 import { getIsTuistInstalled, tuistClean, tuistEdit, tuistGenerate, tuistInstall } from "../common/cli/scripts";
+import type { ExtensionContext } from "../common/commands";
 import { ExtensionError } from "../common/errors";
 
 async function tuistCheckInstalled() {
@@ -10,7 +11,8 @@ async function tuistCheckInstalled() {
   }
 }
 
-export async function tuistGenerateCommand() {
+export async function tuistGenerateCommand(context: ExtensionContext) {
+  context.updateProgressStatus("Running 'tuist generate'");
   await tuistCheckInstalled();
 
   const raw = await tuistGenerate();
@@ -24,7 +26,8 @@ export async function tuistGenerateCommand() {
   vscode.window.showInformationMessage("The Xcode project was successfully generated using Tuist.");
 }
 
-export async function tuistInstallCommand() {
+export async function tuistInstallCommand(context: ExtensionContext) {
+  context.updateProgressStatus("Running 'tuist install'");
   await tuistCheckInstalled();
 
   await tuistInstall();
@@ -34,7 +37,8 @@ export async function tuistInstallCommand() {
   vscode.window.showInformationMessage("The Swift Package was successfully installed using Tuist.");
 }
 
-export async function tuistCleanCommand() {
+export async function tuistCleanCommand(context: ExtensionContext) {
+  context.updateProgressStatus("Running 'tuist clean'");
   await tuistCheckInstalled();
 
   await tuistClean();
@@ -42,7 +46,8 @@ export async function tuistCleanCommand() {
   vscode.window.showInformationMessage("Tuist cleaned.");
 }
 
-export async function tuistEditComnmand() {
+export async function tuistEditComnmand(context: ExtensionContext) {
+  context.updateProgressStatus("Running 'tuist edit'");
   await tuistCheckInstalled();
 
   await tuistEdit();